# Improve binary read performance