This PR makes a couple of improvements to binary read performance:

1. One of the slowest things during a read with many messages is initializing a BinaryReader for each message with its DataView and TextDecoder. This PR adds a `messageField()` method to BinaryReader that reads a size-delimited message field without constructing a new BinaryReader. (`Message.fromBinary` is refactored to share code). A knock-on effect of this change is that we pass `options` along without repeated calls to `makeReadOptions`, which otherwise needlessly spreads `options` into `readDefaults` when `options` is already equivalent to `readDefaults`.
1. readScalar was calling `reader[method]()` which seems to incur an implicit call to .bind() on the reader method (i.e. it is equivalent to `reader[method].bind(reader)()`. By using an inline switch statement we can avoid that indirection and use a faster path.

I put together a simple benchmark (https://github.com/kbongort/protobuf-es/pull/1) using the following message definitions:
```
message User {
  string first_name = 1;
  string last_name = 2;
  bool active = 3;
  User manager = 4;
  repeated string locations = 5;
  map<string, string> projects = 6;
  int32 number = 7;
}

message UserList {
  repeated User users = 1;
}
```

The benchmark uses both a more complete version of the User message and one with only a few fields set:
```
const userMessage = new User({
  firstName: "Jane",
  lastName: "Doe",
  active: true,
  manager: { firstName: "Jane", lastName: "Doe", active: false },
  locations: ["Seattle", "New York", "Tokyo"],
  projects: { foo: "project foo", bar: "project bar" },
});

const smallUserMessage = new User({
  manager: { active: true },
  number: 2,
});
```

After warming up the cache by reading a serialized userMessage 1000 times, it measures the time to read `userMessage`, to read a UserList of 1000 `userMessage`s, and to read a UserList of 1000 `smallUserMessage`s. The amount of time it takes to call `JSON.parse()` on the JSON version of the message is provided as a point of comparison. Without the changes in this PR, the results are:
```
Benchmark: Single user
	Time to parse JSON: 1135 µs
	Time to parse from binary: 3548 µs
	Ratio: 312%
Benchmark: List of 1000 users
	Time to parse JSON: 920 ms
	Time to parse from binary: 2957 ms
	Ratio: 321%
Benchmark: List of 1000 small users
	Time to parse JSON: 260 ms
	Time to parse from binary: 923 ms
	Ratio: 355%
```

After the changes in this PR, the results are:
```
Benchmark: Single user
	Time to parse JSON: 1088 µs
	Time to parse from binary: 2755 µs
	Ratio: 253%
Benchmark: List of 1000 users
	Time to parse JSON: 921 ms
	Time to parse from binary: 1729 ms
	Ratio: 188%
Benchmark: List of 1000 small users
	Time to parse JSON: 255 ms
	Time to parse from binary: 237 ms
	Ratio: 93%
```

The three test cases show speedups of 29%, 71%, and 290% respectively. The effect of (1) and is especially pronounced when there are many small messages, as in the last benchmark test.