@smaye81 2023-05-08
Hi @kbongort appreciate your patience on this. I know we spoke offline last week. We have been gathering some baseline metrics on our performance to see all the areas that need addressed. I can confirm that this does improve binary read performance by an average of ~2.5x.  

However, one concern is the change to modify the `IBinaryReader` interface. This is an interface we expose from the library and changing it will result in breaking changes. Are you up for potentially modifying the PR so that these changes can be made in a non-breaking way without affecting the interface contract?