@kbongort 2023-05-09
> Let me talk it over with the team. We know that the `fromBinary > readMessage` path is a bottleneck, but I'm not sure if there is a good way to reuse the `BinaryReader` like you're doing without breaking one of the interfaces.
> 
> In the meantime, would you mind splitting out the two improvements (`BinaryReader` vs. `readScalar`). I think we can probably get the latter implemented and at least get a quick win.

What type of breakage are you concerned with, exactly? This does add a couple of methods to IBinaryReader. Are you worried that clients may have IBinaryReader implementations that will no longer conform to the interface, or something else?