@kbongort 2023-05-23
Hi Steve,

I believe that I've addressed your concerns with this PR. Can you please
take another look?

Thanks,
Ken

On Tue, May 9, 2023 at 3:13 PM Steve Ayers ***@***.***> wrote:

> Let me talk it over with the team. We know that the fromBinary >
> readMessage path is a bottleneck, but I'm not sure if there is a good way
> to reuse the BinaryReader like you're doing without breaking one of the
> interfaces.
>
> In the meantime, would you mind splitting out the two improvements (
> BinaryReader vs. readScalar). I think we can probably get the latter
> implemented and at least get a quick win.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/bufbuild/protobuf-es/pull/459#issuecomment-1540955378>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/ABGSDEQCO6KDTS7CIJNFVVTXFK6O5ANCNFSM6AAAAAAXAYRMIA>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>